KInda fixed nearly everything
Worked on this for a few hours and didn't quite push the changes up so thats why this is so long. Fixed a lot of issues still a little off of performing as advertised through figma. But a lot of stuff was very unnecessary so I rid of it. One current issues is that its still showing my terminal as 'inuse' so Ill try to fix that; hence I am pushing my changes.